Add 'creation' game lib for creation engine changes, eliminate duplicate plugin code
