+++ CMakeLists.txt
@@ -13,4 +13,5 @@ LIST(APPEND CMAKE_PREFIX_PATH ${QT_ROOT}/lib/cmake)
 LIST(APPEND CMAKE_PREFIX_PATH ${LZ4_ROOT}/dll)
 message(${LZ4_ROOT})
 
-ADD_SUBDIRECTORY(src)
+ADD_SUBDIRECTORY(src/gamebryo)
+ADD_SUBDIRECTORY(src/creation)
