+++ src/gamebryo/gamebryogameplugins.cpp
