+++ src/gamebryo/gamebryogameplugins.h
