+++ src/gamebryo/gamebryosavegameinfowidget.h
