+++ src/gamebryo/gamebryoscriptextender.cpp
