+++ src/gamebryo/gamebryoscriptextender.h
