+++ src/gamebryo/gamebryounmanagedmods.cpp
